32/64-bit fixes.

+++ src/Include/XalanAutoPtr.hpp
@@ -214,7 +214,7 @@ public:
 	}
 
 	Type&
-	operator[](unsigned long	index) const
+	operator[](size_t	index) const
 	{
 		return m_pointer[index];
 	}
